feat: split up into validation_mode_view_model
